+++ prediction_market_agent_tooling/tools/balances.py
@@ -1,12 +1,8 @@
-from typing import Callable
-
 from pydantic import BaseModel
 from web3 import Web3
 from web3.types import Wei
 
-from prediction_market_agent_tooling.config import APIKeys
 from prediction_market_agent_tooling.gtypes import ChecksumAddress, xDai
-from prediction_market_agent_tooling.markets.markets import MarketType
 from prediction_market_agent_tooling.markets.omen.omen_contracts import (
     WrappedxDaiContract,
 )
@@ -29,15 +25,3 @@ def get_balances(address: ChecksumAddress, web3: Web3 | None = None) -> Balances
     xdai = wei_to_xdai(xdai_balance)
     wxdai = wei_to_xdai(WrappedxDaiContract().balanceOf(address, web3=web3))
     return Balances(xdai=xdai, wxdai=wxdai)
-
-
-def get_balance_fn(market_type: MarketType) -> Callable[[], float]:
-    if market_type == MarketType.OMEN:
-        keys = APIKeys()
-
-        def balance_fn() -> float:
-            return float(get_balances(keys.bet_from_address).total)
-
-        return balance_fn
-    else:
-        raise ValueError(f"Unsupported market type: {market_type}")
